+++ src/components/va-charts/VaChart.vue
@@ -5,7 +5,7 @@
 <script setup lang="ts">
   import { computed, ref } from 'vue'
   import type { TChartOptions } from 'vue-chartjs/dist/types'
-  import { defaultConfig, chartTypesMap } from './VaChartConfigs'
+  import { defaultConfig, chartTypesMap } from './vaChartConfigs'
   import { TChartData } from '../../data/types'
 
   const props = defineProps<{
