+++ src/components/va-charts/vaChartConfigs.js
