+++ src/data/charts/index.ts
@@ -1,6 +1,6 @@
-export { bubbleChartData } from './BubbleChartData'
-export { doughnutChartData } from './DoughnutChartData'
-export { barChartData } from './BarChartData'
-export { horizontalBarChartData } from './HorizontalBarChartData'
-export { lineChartData } from './LineChartData'
-export { pieChartData } from './PieChartData'
+export { bubbleChartData } from './bubbleChartData'
+export { doughnutChartData } from './doughnutChartData'
+export { barChartData } from './barChartData'
+export { horizontalBarChartData } from './horizontalBarChartData'
+export { lineChartData } from './lineChartData'
+export { pieChartData } from './pieChartData'
