+++ src/main.ts
@@ -1,10 +1,11 @@
 import { createApp } from 'vue'
-import stores from './stores'
-import router from './router'
 import i18n from './i18n'
 import { createVuestic } from 'vuestic-ui'
-import vuesticGlobalConfig from './services/vuestic-ui/global-config'
 import { createGtm } from '@gtm-support/vue-gtm'
+
+import stores from './stores'
+import router from './router'
+import vuesticGlobalConfig from './services/vuestic-ui/global-config'
 import App from './App.vue'
 
 const app = createApp(App)
